@@ -4,6 +4,9 @@
     "watch": "NODE_ENV=development webpack --config webpack.config.client.js --progress --watch",
     "build:test": "NODE_ENV=development webpack --config webpack.config.client.js --progress --colors",
     "build:prod": "npm run lint && NODE_ENV=production webpack --config webpack.config.client.js --progress --colors -p",
+    "win:watch": "set \"NODE_ENV=development\" && webpack --config webpack.config.client.js --progress --watch",
+    "win:build:test": "set \"NODE_ENV=development\" && webpack --config webpack.config.client.js --progress --colors",
+    "win:build:prod": "npm run lint && set \"NODE_ENV=production\" && webpack --config webpack.config.client.js --progress --colors -p",
     "lint": "eslint --ext .js --ext .jsx ."
   },
   "license": "ISC",
@@ -15,66 +18,70 @@
     "redux-devtools": "^3.3.2",
     "redux-devtools-dock-monitor": "^1.1.1",
     "redux-devtools-log-monitor": "^1.2.0",
-    "webpack-dev-middleware": "1.8.4",
-    "webpack-hot-middleware": "2.13.2"
+    "webpack-dev-middleware": "1.12.1",
+    "webpack-hot-middleware": "2.20.0"
   },
   "dependencies": {
     "assemble": "^0.17.1",
-    "autoprefixer": "^6.5.3",
-    "axios": "0.15.3",
-    "babel-cli": "6.18.0",
-    "babel-core": "^6.18.2",
-    "babel-loader": "6.2.9",
-    "babel-plugin-react-transform": "2.0.2",
-    "babel-preset-es2015": "^6.13.1",
-    "babel-preset-react": "^6.11.1",
+    "autoprefixer": "^6.7.7",
+    "axios": "^0.15.3",
+    "babel-core": "^6.26.0",
+    "babel-eslint": "^8.0.2",
+    "babel-loader": "^7.1.2",
+    "babel-plugin-react-transform": "^2.0.2",
+    "babel-preset-env": "^1.6.1",
+    "babel-preset-react": "6.11.1",
     "babel-preset-stage-0": "6.16.0",
     "body-parser": "1.15.2",
     "bootstrap-css": "^3.0.0",
     "child-process-promise": "2.2.0",
     "classnames": "^2.2.5",
+    "client": "0.0.1",
     "component": "^1.1.0",
     "components": "^0.1.0",
     "containers": "0.0.1",
     "css": "^2.2.1",
-    "css-loader": "^0.26.1",
+    "css-loader": "^0.28.7",
     "css-raw-loader": "^0.1.1",
+    "cssnano": "^3.10.0",
     "eslint": "3.15.0",
     "eslint-config-airbnb": "12.0.0",
     "eslint-loader": "1.6.1",
-    "eslint-plugin-import": "2.2.0",
+    "eslint-plugin-import": "^2.8.0",
     "eslint-plugin-jsx-a11y": "2.2.3",
     "eslint-plugin-react": "6.9.0",
-    "extract-text-webpack-plugin": "^1.0.1",
+    "extract-text-webpack-plugin": "^2.1.2",
     "file-loader": "^0.9.0",
-    "fontgen-loader": "^0.2.1",
     "helpers": "0.0.6",
+    "index": "^0.4.0",
     "json-loader": "0.5.4",
-    "less": "^2.7.1",
-    "less-loader": "^2.2.3",
-    "less-plugin-glob": "^1.1.1",
+    "loader-utils": "^1.1.0",
     "local-storage": "^1.4.2",
     "lodash": "^4.17.4",
     "npm-install-webpack-plugin": "^4.0.4",
     "object-assign": "^4.1.0",
     "postcss-cssnext": "^2.9.0",
+    "postcss-flexbugs-fixes": "^3.2.0",
     "postcss-import": "^9.0.0",
     "postcss-loader": "^1.2.0",
     "postcss-reporter": "^2.0.0",
-    "react": "15.4.1",
-    "react-dom": "15.4.1",
-    "react-redux": "4.4.6",
-    "react-router": "3.0.0",
-    "react-router-redux": "4.0.7",
+    "prop-types": "^15.6.0",
+    "react": "16.1.1",
+    "react-dom": "16.1.1",
+    "react-redux": "5.0.6",
+    "react-router": "^3.2.0",
+    "react-router-redux": "4.0.8",
     "reducers": "^3.0.0-alpha",
-    "redux": "3.6.0",
-    "redux-logger": "2.7.4",
+    "redux": "3.7.2",
+    "redux-logger": "3.0.6",
     "redux-thunk": "2.1.0",
-    "scss": "^0.2.4",
-    "style-loader": "^0.13.1",
+    "style-loader": "^0.19.0",
     "styles": "^0.2.1",
+    "svg-sprite-loader": "^3.4.1",
+    "svgo": "^1.0.3",
+    "svgo-loader": "^2.1.0",
     "swiper": "^3.4.1",
     "url-loader": "^0.5.7",
-    "webpack": "^1.13.2"
+    "webpack": "^2.6.1"
   }
 }
