@@ -1,8 +1,10 @@
 /* eslint react/prefer-stateless-function: "off" */
 /* Because stateless functions don't have context it seems */
 import React from 'react';
+import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import strings from './strings';
+
+import strings from '../json/content.json';
 
 export default function translate(key) {
     return Component => {
@@ -35,8 +37,8 @@ export default function translate(key) {
         }
 
         TranslationComponent.propTypes = {
-            currentStrings: React.PropTypes.object,
-            currentLanguage: React.PropTypes.string
+            currentStrings: PropTypes.object,
+            currentLanguage: PropTypes.string
         };
 
         return connect(stateToProps)(TranslationComponent);
