@@ -2,11 +2,7 @@ import React from 'react';
 import IndexHero from '../PageIndexHero';
 import IndexPartners from '../PageIndexPartners';
 
-const Index = () => (
-    <div>
-        <IndexHero />
-        <IndexPartners />
-    </div>
-);
-
-export default Index;
+export default () => [
+    <IndexHero key="index-hero" />,
+    <IndexPartners key="index-partners" />
+];
