@@ -1,13 +1,11 @@
-import React, { PropTypes } from 'react';
-import objectAssign from 'object-assign';
+import React from 'react';
+import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
-import { grid } from 'bootstrap-css';
 
 import Animate from '../Animate';
 import styles from './styles.css';
 import translate from '../../i18n/Translate';
 
-objectAssign(styles, grid);
 const cx = classNames.bind(styles);
 
 const IndexAbout = (props) => (
