@@ -1,49 +1,49 @@
-import React, { PropTypes } from 'react';
-import objectAssign from 'object-assign';
+import React from 'react';
+import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
-import { grid, type } from 'bootstrap-css';
 
 import Animate from '../Animate';
 import styles from './styles.css';
 import translate from '../../i18n/Translate';
 
-objectAssign(styles, type, grid);
 const cx = classNames.bind(styles);
 
 const IndexPartners = (props) => (
     <div className={cx('partners')}>
         <Animate type="bottom-to-top" startFrom={1300}>
-            <h3 className={cx('text-center')}>{ props.strings.trusted }</h3>
+            <div className={cx('text-center', 'partners-title')}>{ props.strings.trusted }</div>
         </Animate>
-        <div className={cx('partner-logo')} >
-            <Animate type="bottom-to-top" startFrom={1300}>
-                <img src="/app/i/partners/articulmedia.png" alt="Articul Media" className={cx('articul')} />
-            </Animate>
-        </div>
-        <div className={cx('partner-logo')} >
-            <Animate type="bottom-to-top" startFrom={1350}>
-                <img src="/app/i/partners/cpeople.png" alt="CreativePeople" className={cx('cpeople')} />
-            </Animate>
-        </div>
-        <div className={cx('partner-logo')} >
-            <Animate type="bottom-to-top" startFrom={1400}>
-                <img src="/app/i/partners/notamedia.png" alt="Notamedia" className={cx('nota')} />
-            </Animate>
-        </div>
-        <div className={cx('partner-logo')} >
-            <Animate type="bottom-to-top" startFrom={1450}>
-                <img src="/app/i/partners/npfsb.png" alt="НПФ Сбербанка" className={cx('npfsb')} />
-            </Animate>
-        </div>
-        <div className={cx('partner-logo')} >
-            <Animate type="bottom-to-top" startFrom={1500}>
-                <img src="/app/i/partners/yota.png" alt="Yota" className={cx('yota')} />
-            </Animate>
-        </div>
-        <div className={cx('partner-logo')} >
-            <Animate type="bottom-to-top" startFrom={1550}>
-                <img src="/app/i/partners/zephyr.png" alt="Zephyrlab" className={cx('zephyr')} />
-            </Animate>
+        <div className={cx('partners-list')}>
+            <div className={cx('partner-logo')} >
+                <Animate type="bottom-to-top" startFrom={1300}>
+                    <img src="/app/i/partners/articulmedia.png" alt="Articul Media" className={cx('articul')} />
+                </Animate>
+            </div>
+            <div className={cx('partner-logo')} >
+                <Animate type="bottom-to-top" startFrom={1350}>
+                    <img src="/app/i/partners/cpeople.png" alt="CreativePeople" className={cx('cpeople')} />
+                </Animate>
+            </div>
+            <div className={cx('partner-logo')} >
+                <Animate type="bottom-to-top" startFrom={1400}>
+                    <img src="/app/i/partners/yota.png" alt="Yota" className={cx('yota')} />
+                </Animate>
+            </div>
+            <div className={cx('partner-logo')} >
+                <Animate type="bottom-to-top" startFrom={1450}>
+                    <img src="/app/i/partners/notamedia.png" alt="Notamedia" className={cx('nota')} />
+                </Animate>
+            </div>
+            <div className={cx('partner-logo')} >
+                <Animate type="bottom-to-top" startFrom={1500}>
+                    <img src="/app/i/partners/npfsb.png" alt="НПФ Сбербанка" className={cx('npfsb')} />
+                </Animate>
+            </div>
+            <div className={cx('partner-logo')} >
+                <Animate type="bottom-to-top" startFrom={1550}>
+                    <img src="/app/i/partners/zephyr.png" alt="Zephyrlab" className={cx('zephyr')} />
+                </Animate>
+            </div>
         </div>
     </div>
 );
