@@ -7,18 +7,19 @@ import { createStore, compose, applyMiddleware } from 'redux';
 import { Provider } from 'react-redux';
 import { Router, browserHistory } from 'react-router';
 import { syncHistoryWithStore } from 'react-router-redux';
+import { createLogger } from 'redux-logger';
 import createRoutes from './routes';
 import reducers from './state';
 import DevTools from './components/DevTools';
+import './sprite';
 
 const initialState = JSON.parse(document.getElementById('app').getAttribute('data-state'));
 
 let store;
 if (__DEVMODE__) {
-    const createLogger = require('redux-logger');
     const logger = createLogger({ predicate:
         (getState, action) => action.type !== 'EFFECT_TRIGGERED' &&
-        action.type !== 'EFFECT_RESOLVED' });
+                              action.type !== 'EFFECT_RESOLVED' });
     const enhance = compose(applyMiddleware(logger), DevTools.instrument());
 
     store = createStore(reducers, initialState, enhance);
@@ -34,12 +35,10 @@ let rootElements;
 if (__DEVMODE__) {
     rootElements = (
         <Provider store={store}>
-            <div>
-                <Router history={history}>
-                    {routes}
-                </Router>
+            <Router history={history}>
+                {routes}
                 <DevTools />
-            </div>
+            </Router>
         </Provider>
     );
 } else {
